[SDK] test: Add e2e test for tune function. (#2399)
* fix(sdk): fix error field metrics_collector in tune function.


* test(sdk): Add e2e tests for tune function.


* test(sdk): add missing field parameters.


* refactor(test/sdk): add run-e2e-tune-api.py.


* test(sdk): delete tune testing code in run-e2e-experiment.


* test(sdk): add blank lines.


* test(sdk): add verbose and temporarily delete e2e-experiment test.


* test(sdk): add namespace_labels.


* test(sdk): add time.sleep(5).


* test(sdk): add error output.


* test(sdk): build random image for tune.


* test(sdk): delete extra debug log.


* refactor(test/sdk): create separate workflow for tune.


* test(sdk): change api to API.


* test(sdk): change the permission of scripts.


* test(sdk): delete exit code & comment image pulling.


* test(sdk): delete image pulling phase.


* test(sdk): refactor workflow file to use template.


* test(sdk): mark experiments and trial-images as not required.


* test(sdk): pass tune-api param to setup-minikube.sh.


* test(sdk): fix err in template-e2e-test.


* test(sdk): add debug logs.


* test(sdk): reorder params and delete logs.


---------
